+++ components/retro-go/fonts/fonts.h
@@ -3,6 +3,10 @@
 /**
  * This file can be edited to add fonts to retro-go.
  * To create new fonts you can use font_converter.py located in the tools folder.
+ *
+ * The Sans fonts carry the Arabic block and the Arabic presentation forms on top of Latin-1, which
+ * is what rg_text_shape() needs in order to draw Persian and Arabic joined and right to left. The
+ * other fonts are Latin-only, so pick a Sans one for those languages.
  */
 
 extern const rg_font_t font_basic8x8;
@@ -10,7 +14,8 @@ extern const rg_font_t font_DejaVu12;
 extern const rg_font_t font_DejaVu15;
 extern const rg_font_t font_VeraBold11;
 extern const rg_font_t font_VeraBold14;
-extern const rg_font_t font_MediaSans15;
+extern const rg_font_t font_Sans12;
+extern const rg_font_t font_Sans15;
 
 enum {
     RG_FONT_BASIC_8,
@@ -20,7 +25,8 @@ enum {
     RG_FONT_DEJAVU_15,
     RG_FONT_VERA_11,
     RG_FONT_VERA_14,
-    RG_FONT_MEDIA,
+    RG_FONT_SANS_12,
+    RG_FONT_SANS_15,
     RG_FONT_MAX,
 };
 
@@ -32,5 +38,6 @@ static const rg_font_t *fonts[RG_FONT_MAX] = {
     &font_DejaVu15,
     &font_VeraBold11,
     &font_VeraBold14,
-    &font_MediaSans15,
+    &font_Sans12,
+    &font_Sans15,
 };